+++ scripts/github_release.rb
@@ -1,32 +1,57 @@
 #!/usr/bin/env ruby
 
 require 'fileutils'
+require 'octokit'
 require 'pathname'
 require 'tmpdir'
 
-require 'octokit'
+raise 'usage: github_release.rb version xcode_version' unless ARGV.length == 2
+
+VERSION = ARGV[0]
+XCODE_VERSION = ARGV[1]
+ACCESS_TOKEN = ENV['GITHUB_ACCESS_TOKEN']
+raise 'GITHUB_ACCESS_TOKEN must be set to create GitHub releases' unless ACCESS_TOKEN
 
 BUILD_SH = Pathname(__FILE__).+('../../build.sh').expand_path
-VERSION = `sh '#{BUILD_SH}' get-version`.strip
 RELEASE = "v#{VERSION}"
 
 BUILD = BUILD_SH.parent + 'build'
-OBJC_ZIP = BUILD + "realm-objc-#{VERSION}.zip"
 SWIFT_ZIP = BUILD + "realm-swift-#{VERSION}.zip"
-CARTHAGE_ZIP = BUILD + 'Carthage.framework.zip'
-CARTHAGE_XCFRAMEWORK_ZIP = BUILD + 'Carthage.xcframework.zip'
 
 REPOSITORY = 'realm/realm-swift'
 
-puts 'Creating Carthage XCFramework package'
-FileUtils.rm_f CARTHAGE_XCFRAMEWORK_ZIP
-CARTHAGE_XCODE_VERSION = BUILD_SH.parent.+('Jenkinsfile.releasability').read()[/carthageXcodeVersion = '([0-9.]+)'/, 1]
+$uploads = [SWIFT_ZIP]
+
+def sh(*args)
+  system(*args, :out=>"/dev/null") || exit(1)
+end
+
+def zip(name, *files)
+  path = (BUILD + name).to_path
+  FileUtils.rm_f path
+  sh 'zip', '--symlinks', '-r', path, *files
+  $uploads.append Pathname(path)
+end
 
 Dir.mktmpdir do |tmp|
   Dir.chdir(tmp) do
-    system('unzip', SWIFT_ZIP.to_path, "realm-swift-#{VERSION}/#{CARTHAGE_XCODE_VERSION}/*.xcframework/*", :out=>"/dev/null") || exit(1)
-    Dir.chdir("realm-swift-#{VERSION}/#{CARTHAGE_XCODE_VERSION}") do
-      system('zip', '--symlinks', '-r', CARTHAGE_XCFRAMEWORK_ZIP.to_path, 'Realm.xcframework', 'RealmSwift.xcframework', :out=>"/dev/null") || exit(1)
+    sh 'unzip', SWIFT_ZIP.to_path,
+           "realm-swift-#{VERSION}/*/RealmSwift.xcframework/*",
+           "realm-swift-#{VERSION}/Realm.xcframework/*"
+    Dir.chdir "realm-swift-#{VERSION}" do
+      zip 'Realm.xcframework.zip', 'Realm.xcframework'
+      Dir.glob '*/RealmSwift.xcframework' do |name|
+        version = Pathname(name).parent
+        puts "Creating SPM package for #{version}"
+        Dir.chdir version do
+          zip "RealmSwift@#{version}.xcframework.zip", 'RealmSwift.xcframework'
+        end
+      end
+
+      puts "Creating Carthage package for #{XCODE_VERSION}"
+      FileUtils.mv "#{XCODE_VERSION}/RealmSwift.xcframework",
+                   'RealmSwift.xcframework'
+      zip 'Carthage.xcframework.zip', 'Realm.xcframework', 'RealmSwift.xcframework'
     end
   end
 end
@@ -52,12 +77,11 @@ github = Octokit::Client.new
 github.access_token = ENV['GITHUB_ACCESS_TOKEN']
 
 puts 'Creating GitHub release'
-prerelease = (VERSION =~ /alpha|beta|rc/) ? true : false
+prerelease = (VERSION =~ /alpha|beta|rc|preview/) ? true : false
 response = github.create_release(REPOSITORY, RELEASE, name: RELEASE, body: RELEASE_NOTES, prerelease: prerelease)
 release_url = response[:url]
 
-uploads = [OBJC_ZIP, SWIFT_ZIP, CARTHAGE_ZIP, CARTHAGE_XCFRAMEWORK_ZIP]
-uploads.each do |upload|
+$uploads.each do |upload|
   puts "Uploading #{upload.basename} to GitHub"
   github.upload_asset(release_url, upload.to_path, content_type: 'application/zip')
 end